Include full buildable sources in the main package
This allows the packaged version to be used for debugging as well
as in-place fixing if necessary while a proper PR is merged.
GitInfo.txt for the built version is included too, so that the built
assembly has the exact same version and token as the one shipped
with the package.