Added build-time generation of [Export] and [PartCreationPolicy]
The targets/tasks take care of inspecting the source in an efficient
manner, looking for [Component] attributes and generating partial
classes for them containing the required MEF attributes so that the
nicety of using [Component] doesn't have a negative impact in
performance.

Unfortunately, [ImportingConstructor] will still be needed since
we're not modifying the user's code and that can't be made partial.