Allow skipping namespaces when generating exports
Previously we would generate exports for all interfaces.
By allowing authors to specify an ExcludeInterfaceNamespace
item via MSBuild, they can exclude entire namespaces from
automatic export behavior. By default, Clide excludes
Microsoft.VisualStudio namespace.

An easy way to "opt-in" to a specific interface in such a
namespace is to simply add a manual [Export] with that
interface to the component class, since it's a partial and
the auto-exporting is merely a convenience and all supported
MEF attribute are still available to the component author.