Fixed FilteringProjectsVisitor to avoid throwing an exception for generic nodes
Returning false when we want to stop visiting nodes instead of throwing
an NotImplementedException.

When a project is unloaded in a solution, the IGenericNode is visited.
With the previous implementation we were throwing an exception instead
of just stopping the node visiting at the moment.

There is no advantage of throwing when the node is not supposed to
be visited. We just need to stop continue visiting the children at
that moment.

Also removing the ProjectsVisitor and use the FilteringProjectVisitor instead
with a default predicate (which will return all projects in the solution)