Fixed the Projects visitor to not fail when there are unloaded Projects
When a project is unloaded, the GenericNode is visited so we need to keep
continue visiting nodes at this stage in order to visit the other projects
nodes that come after the unloaded projects. Otherwise once we find an unloaded
project, we stop visiting nodes and some projects are not returned.