Fixed VsHierarchyItemManagerProvider to use AsyncLazy<T> instead of Lazy<T>
Using Lazy<T> with a value factory that makes UI thread switching or use the JoinableTaskContext to do synchronous blocking is considered a threading anti pattern given that it can lead to deadlocks.
For the case of the VsHierarchyItemManagerProvider, we are doing the two things. First, using JoinableTaskFactory to do a synchronous block in order to allow switching to the UI thread inside.

There is already a threading analyzer that warns if this pattern is violated, and explains the different situations: https://github.com/Microsoft/vs-threading/blob/master/doc/analyzers/VSTHRD011.md
This problem has been found after deeply analyzing the logs and dumps of a VS Feedback issue with serious locks in VS when opening Xamarin iOS solutions. This is the feedback issue: https://developercommunity.visualstudio.com/content/problem/289617/opening-xamarin-ios-solution-locks-vs-2017-hard.html

This pattern should be fixed properly for Dev16, together with a complete async patterns review, in all the places and not only in this class and repo, but for d15.9 we are only including this fix to limit the impact.