Added support to avoid Executing commands based on a given UI Context
Added an extra and optional argument to the CommandAttribute to define
the associated UI Context that should be active for the command
to be executed.

When the UI Context is not active neither the QueryStatus or the Execute
should be executed. And also the command should be disabled and invisible.