Using Hierarchy will get all the capabilities instead of getting only the ones from the inner hierarchy.
Fixed creation of project nodes for flavored projects

We must not call Get IVsHierarchyItemManager.GetHierarchyItem if the
corresponding IVsHierarchy does not implement IVsProject. For the case
of flavored project it might or it might not be the case. So we should
always use the inner hierarchy and keep the flavored hierarchy instance
for IVsHierarchy.GetProp scenarios where we should able to:

projectNode.AsVsHierarchy().GetProperty(x)

where x is provided by the flavored project so AsVsHierarchy() should
retrieve the flavored instance.