Fix CI build dependencies
Build against a fixed version of the XVSSDK in CI, instead of a floating
version, since that can introduce unwanted dependency changes when building the
same commit (or a later one). The version specified is the one we shipped with
16.0.

Also, don't append branch name for v* branches (they are release branches). This
simplifies the nuget package build metadata, no actual versioning change.