Remove unnecessary SVsAppId
In some cases, the SVs/IVs pattern is used to assign a service id (The SVs interface’s GUID) that differs from the ids of any interfaces it may happen to implement. Some times, it is just aliased to be the same id.

In this second case, declaring a dummy type with the same Guid is *not* needed, and it can moreover cause issues in the CLR as two COM types will end up having the same ID.

In service queries the only thing that matters is the GUIDs not the type, so for the purposes of GetService the types given are just a place to hang a GUID.

This commit removes the SVsAppId and replaces it with IVsAppId when queried.