Improved ReferencesNodeFactory for non-english loc based heuristic
For non english loc based the factory was failing because the HierarchyItem.Text is translated
to the target language.

So we're now improving the logic by also checking if any child is a VSLanhProj.Reference

We went through all the __VSHPROPID values and we did not find anything usefull to determine
if the given Hierarchy was indeed a ReferencesNode.