Limit minimum VS version to current in Debug builds
When building in debug mode, we use the current VS version to limit the target VS and max/min versions, so that tests only run against the current VS version. We also disable recycling since in debug builds, we'll most likely want to see tests fail once and fix them.