Fixed the Nodes traversal
The HierarchyManager was not implemented to be thread safe.
If the Nodes are being traversed from a background thread,
the caller might end up with a duplicate key/item exception.

So the proposal fix is to replace the existing implementation
by traversing the hierarchy using the legacy FirstChild/NextSibling
mechanism and switch to the main thread only when we need to
get/create the HierarchyItem.