Ensure building on Mac and Win succeed
All projects that specify a $(TargetOS) that matches the current $(OS) will be built.
Otherwise, the NoBuild.targets is imported, which overrides Build and CoreBuild to
not build anything. We also condition the project references in the VSIX to the same
effect

The goal is twofold:
1. If the solution is being built on the Mac, only the Mac projects and mpack should be
   built
2. If the solution is being built on Windows, we should build the Win projects and VSIX.

This makes the developer experience intuitive for working on both platforms easily.

Instead of attempting to create a VSM-compatible addin from the Vsix project,
we just have another "head" for VSM, and we just conditionally build Mac or Win projects.
Note that we don't need to keep special solution configs or even two solutions since we
just disable the build via the `TargetOS` property.