Adjust checkout hash before upload
When the current commit from the checkout doesn't match the GH commit to report to, force a reset to that commit so the artifacts contain the right hashes.

This works in coordination with the v2 template for fixing the source version, which sets the GH.Commit variable we use afterwards.