Automatically detect the presence of an experimental Clide
Whenever Clide is built from source, it will be deployed and enabled in
the experimental instance of VS. Since the install path in the experimental
instance is predictable, we can automatically attempt to find it and offer
a better inner loop for contributors by replacing the nuget-provided Clide
references in their VSIX with the ones deployed by a locally build Clide.

This means contributing and consuming the contribution is as easy as
building Clide and then build as normal your extension. At some point when
the new APIs are available via nuget, you just bump the package reference
and can remove the experimental Clide from the local deploy folder at
`%LocalAppData%\Microsoft\VisualStudio\16.0_[INSTANCEID]Exp\Extensions\ClariusLabs\Clide\42.42.42`.
A quick `devenv /updateConfiguration /rootSuffix Exp` from the developer
command prompt would quickly unregister it and the newly updated/inserted
Clide (from the main VS) will again be the one in use.